Change DB healthcheck to just use 127.0.0.1
Before changing this the server_container would never start because the
db_container never reached a "healthy" status. I discovered this was
happening because in my local environment running hostname -i within the
docker container outputs many IP addresses, not one, and results in an
invalid posgtres URL.

I think this may be because I have configured multiple docker bridge
networks in the past. Ulitmately hostname -i does not seem to be a good
choice here because it does not consistently output a single IP across platforms
and environments.